core: Disallow RO disks to be attached to VM
RO is not supported when using specific inrtefaces when attached to a
VM. These case are already disallowed when creating a disk, yet are
still allowed when attaching a floating disk with these interfaces, or
when updating a disk.
These interface are IDE and iSCSI for a LUN target. This patch adds the
relevant validations to AttachDiskToVmCommand and UpdateVmDiskCommand's CDAs.

Change-Id: Iac2a8e0fbece57cbb6f375a65cc37e7c2bba11cf
Bug-Url: https://bugzilla.redhat.com/1082673
Bug-Url: https://bugzilla.redhat.com/1094288
Signed-off-by: Vered Volansky <vvolansk@redhat.com>