aaa: add authz namespace support
in an environment of large network as active directory forest we need a
way to query subset of the content. the namespace will enable the
provider to place partitions / domains into context to allow user
interface to present selector.

currently user interface does not support that, so we query all
namespaces at server side.

Topic: AAA
Change-Id: I38968ae4b61aa8d0a5120a22a17e016cbe96c5e9
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>