packaging: setup: export iso domain to engine only
Since Ia35aaf42cb46247a2a02df1c2418f4613ad6b965 we ask the user for the
ACL to use when exporting the iso domain. We kept the default to be the
same as the old behavior - export read-write to the entire network.

This is a security risk, potentially allowing unauthorized changes to
iso images.

This change makes the iso domain accessible by default to the engine's
machine only.

Change-Id: I504d608896277d66b7c8adfe6e7f45ea00255fb8
Signed-off-by: Yedidyah Bar David <didi@redhat.com>