core: copy disks from template instead of instance type
When the VM was created from both template and instance type, all the devices
(including the disks) were taken from the instance type.

The problem is that the disks are supposed to be taken from the template.
This patch ensures that if the VM is created from an instance type, the disks
are taken from the template.

Change-Id: Iaddc4d3da99c0a459c7339639cca15cbb597df18
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>