core: add time tolerance to websocket proxy
This small patch adds 5 second tolerance to the past in
ovirt-websocket-proxy. This is to prevent really small time drifts
between clocks of the engine and the machine where the proxy is deployed
in situations when the proxy clock is ahead of the engine clock (and
therefore the ticket is not yet valid from the proxy POV).

Change-Id: I811f6c3c4413576c1dc8d209408767367b3687ad
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1052921