core: New internal command, VmSlaPolicyCommand
Adding a new command to handle the SLA parameters.
For now the new command only contains the CPU Limit parameter
which limits the VM's CPU consumption.

This Command will be called from the updateVmCommand, and updateMomPolicyCommand.
later on it will be called from other parts of the system that configures SLA parameters.

Change-Id: I26b0794c52cfe1d352f26392a835023afcd9efa0
Bug-Url: https://bugzilla.redhat.com/1084930
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>