webadmin: Vlan id should be shown next to vlan device name
In order to show the vlan id next to the vlan device name some changes
were done to the host interface sub tab.
1. Removing the address from the brackets next to the vlan device and bond
name.
2. Moving the adress column to the VLanPanel next to the network column.
All the addresses, vlan and non-vlan will be shown in this column.
3. Splitting the Interface panel into two panels- InterfacePanel and
StatisticsPanel. The StatisticsPanel will be the last panel- after the
VLanPanel.

Notice: there were some look-and-feel issues caused to this tab in commit
0248286a6b70ffe9076731aea406fdbfd6a97b0a.
This patch doesn't fix them.

Change-Id: Ia900e7d293df479cd9c98cfe66dcd0cb8b1f4fe4
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>