webadmin: Repair disallow RO in UI
Qemu currently does not support direct-LUN disks connected using
VirtIO-SCSI. This patch blocks this option from UI when
adding/editing/attaching a direct-LUN disk with VirtIO-SCSI interface to a VM.
This wasn't checked in Add/Edit Disk to Vm dialog, nor in Attach table.
IDE behaviour was changed to comply with this behaviour in this patch as well.
This patch also removes un-needed verfication for IDE disk.

Change-Id: Ieb1923d03785be7f228076e12f4865242aed5c90
Bug-Url: https://bugzilla.redhat.com/1097754
Signed-off-by: Vered Volansky <vvolansk@redhat.com>