setup: Write the answer file if file path is valid
Added a exception handling to make sure answer file gets
written only if the file path is valid one. If the path
is invalid, instead of erroring out and marking the whole
setup as failed, it shows a warning and setup is marked as
completed successfully.

A valid use case is calling "engine-setup --generate-answer
=root/engine-answer.txt". In this case the current working
dir does not contain a directory with name root.

In the above scenario, earlier engine setup used to display an
error saying "No such file or directory: '<CurrDir>/root/
engine-answer.txt'". Also the overall engine-setup output used
to report "[ ERROR ] Execution of setup failed".

With the fix, now the output looks as below -
[WARNING] [Errno 2] No such file or directory:
'<CurrDir>/root/engine-answer.txt'
and the overall engine setup reports a success as below -
"[ INFO  ] Execution of setup completed successfully"

Change-Id: I8c595449e631f0bdb9524d74bdc3d9b1a256439f
Bug-Url: https://bugzilla.redhat.com/1089067
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>