webadmin: Simplified failure handling code in Frontend
Changed the name of a variable to better represent what it means, and
extracted some common code out of conditional clauses.

This also exposed some dead code - there was some logic setting an
alternative error message in case the failure isn't on canDoAction(),
but it was located in a conditional clause where it had already been
established that the failure had been on canDoAction().

Also took the chance to remove a redundant event flow, where a single
error message was being passed around instead of a list, but which in
the end would perform the same logic.

Change-Id: I19a0fc10841c9d400ad381f5f9f5fbaaf55fe871
Signed-off-by: Lior Vernia <lvernia@redhat.com>