engine: Simplify code in UpdateVdsGroupCommand
Removed a condition that's blocked by canDoAction() - moving a cluster
from one (non-null) DC to another. The realization that this operation
is blocked also enables to simplify the management network logic, and
to always add it and assign it all roles (if it weren't blocked, then
the current code wouldn't be correct anyway, because other networks
could occupy the roles).

Change-Id: I0091d9afb6d4b26c4ba9fb434192a56eda9397e4
Signed-off-by: Lior Vernia <lvernia@redhat.com>