extmgr: use extension class loader when calling extension
currently the context class loader is set to the jboss module that is at
top level. this does not allow the extension to use extra dependencies
it may have in its module.xml.

to solve this we set the context class loader of the extension module
before calling the extension, in this case the extension finds all its
dependencies.

Change-Id: I38031776f1d740da8f8f07f0d4a52b640534bcff
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>