restapi: Accept Expect and X-Ovirt-Expect
The HTTP Expect header is rejected by the Apache web server if its value
is anything other than 100-continue. To avoid problems we have an
explicit configuration to remove it. As users may want to use that
behaviour this patch changes the RESTAPI so that it will accept both the
Expect and X-Ovirt-Expect headers.

Change-Id: I6b66b1f04b286c756d4c1e89c3555c1c95593881
Bug-Url: https://bugzilla.redhat.com/1093784
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>