core: removing lock from DetachStorageDomainFromPoolCommand
The lock on the domain id when detaching it was added in change
I4ee8cd4b64393d195fe45349f49996e45ce09f0f but it's currently incorrect
because the detach command is also executed from commands that already
take the lock on the storage domain, which will lead to a failure.
We need to pass the command context to avoid that situation, but that
will be inspected in a further patch - right now this patch returns the
situation to it's previous state.

Change-Id: I90b801416bb932ddc08dbb2d626882fc60895826
Signed-off-by: Liron Aravot <laravot@redhat.com>