engine : Introduce TimeConverter
This patch introduces TimeUnitConverter that converts
(approximately) the input difference in times(long)
to a maximum expressable time unit.Currently it goes
upto days.
This patch is just a wrapper around the java concurrent
TimeUnit.The wrapper is written with the intention of
providing the additional facility of auto-converting
the lower unit time/time interval into its highest
possible expressible unit such that its value post
conversion > 1
Also,this patch just returns the integral magnitude and
unit of the highest expressable unit.

Change-Id: I045aaa4f93b345e0457e1a0ce2ccf0ceeb9071ad
Signed-off-by: Anmol Babu <anbabu@redhat.com>