engine: UserRole permission on System/DC/Cluster doesn't work as expected
When granting UserRole permission on System/DC/Cluster,
it allows the user to see all the VMs under that entity
(System/DC/Cluster), including Pooled VMs.  The result is
that he will see all the VMs (including the Pooled VM)
in the User Portal, but he won't see the VM Pools
(as the logic is to show only pools that the user has
no UserRole on at least one VM from this pool).

Change-Id: If106d64939b80de65fb7052e97f69a6d5de0b010
Bug-Url: https://bugzilla.redhat.com/863074
Signed-off-by: Ravi Nori <rnori@redhat.com>