packaging: spec: support distro without jboss-as 7
A new package providing jboss-as 7 for distro that
don't ship it has been created as ovirt-engine-jboss-as.
The package can be installed side-by-side with wildfly on
Fedora 20 and doesn't conflict with jboss-as manual installations
on /usr/share/jboss-as using its own directory.
The package is built by
http://jenkins.ovirt.org/job/ovirt-engine-jboss-as_master_create-rpms_merged

The spec file has been updated for EL6 and F20 in order to require
the new package and provide the environment keys for automatically
set jboss home directory.

Change-Id: Icef92a1f22e6f88458d1624429dff57a9817192f
Bug-Url: https://bugzilla.redhat.com/1055604
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>