engine: errors in audit log- VDS_SET_NONOPERATIONAL_IFACE_DOWN
The format of the log was-
"Host 'host-name' moved to Non-Operational state because interfaces
'iface-list are down but are needed by networks 'net-list' in the current
cluster"
There were some errors in it-
1. Displayed vlan device names (should display just base interface names).
2. Didn't display all the problamatics networks of a bond.
3. Didn't display problematic bonds if there is another problematic
nic.

This patch fixes these errors and also slightly changes the format of the
error message.

Since the bond status can be determined from the bond itself and shouldn't
be calculated from the slaves status and most of the bugs were part of this
logic, this patch removes it.

Change-Id: I9dfa5396c7622b5e689ee3e648fcccd0e75ed834
Bug-Url: https://bugzilla.redhat.com/1079964
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>