webadmin: Aggregate errors by default in runMultipleActions()
Changed the default calls to aggregate errors. Although this
potentially affects several existing invocations, most of them provide
a failure callback of their own and so render this argument
irrelevant.

The ones that are affected are SanStorageModelBase.connectTargets()
and NewDiskModel.onAttachDisks(), where it seemed to me that the
desireable behavior would indeed be to aggregate multiple errors if
they are returned.

Also corrected the Javadoc for some overloads in Frontend.

Change-Id: I423a737888a3b3ec6bb492d8eaf825d9cce9872c
Signed-off-by: Lior Vernia <lvernia@redhat.com>