core: start PM enabled hosts after engine restart
This patch handles the following scenario
host non-responding->Restart->Stop->off->engine restart
In this case the host will not be started again after the engine is up.
This patch handle such scenarios by searching for hosts that both have
PM enabled and are in 'Reboot' status after the
DisableFenceAtStartupInSec quite time in which PM operations are
skipped.
If such hosts are found after the quite time, an attempt to start the
host via its PM card is executed.
Since the status of a Host is moved to 'Reboot' when it is restarted
just after the Stop command (Reboot=stop->wait for off->start->wait for
on), if the engine fails or stopped at this point the Host will be left
in the Reboot status.

Change-Id: Ibbfe83535364556df72eaac348e58d4b7146156e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1005756
Signed-off-by: Eli Mesika <emesika@redhat.com>