core: do not handle crashed VM twice
We used to save the VMs which moves to Down in a List which means that
theoretically we can have duplicated VMs in the List. Because of bz
1099846, we might get into a situation where VM will be in the List
twice.

Before extracting the handling of the VMs that went down to separate
thread, it is important to make sure we have each VM stored in this
collection only one, so we won't invoke redundant threads. So the
solution is to change the List to be Set, so it will not store
duplicated values.

Change-Id: I913cf18ad341090f3ed1df5acda93b372af7ab4d
Bug-Url: https://bugzilla.redhat.com/1108675
Signed-off-by: Arik Hadas <ahadas@redhat.com>