userportal,webadmin: Remove redundant scroll bar in VM general tab
The ProfilesInstanceTypeEditor had the vertical size hardcoded and was wrapped
by a ScrollPanel. This caused the redundant scroll bar in the VM general tab to
appear.
Fixed by removing the ScrollPanel since in all other tabs in the VM dialog the
scrolling is handled by the dialog itself - this change makes the behavior
consistent with the rest of the tabs.

It has been already fixed in http://gerrit.ovirt.org/#/c/26169 but has been
accidentally introduced back in http://gerrit.ovirt.org/#/c/24594 so this
patch removes it again.

Change-Id: I03f31624feaa23b37006310d365d57e717d4bbb0
Bug-Url: https://bugzilla.redhat.com/1047883
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>