core: use Locked status when activating master domain
When we try to activate the master domain we should use the Locked
status because there is no "Activating" transition (it is active by
definition).
In general the master domain is always active, but this special flow
is used on the creation of the StoragePool when the first domain is
added.

Change-Id: I2f398a8fb946b339786f9776f460eab2f920c7b1
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>