webadmin: add buttons in the new cluster dialog and make it closeable
When create a cluster from the system tree panel,
e.g. select a ovirtmgmt node in the system tree
and select Clusters source tab, then click the New button,
the "New Cluster" dialog has no OK button, Cancel button
and cannot be closed.

The selected node in the system tree is not a Data Center
Node, but it was converted into a Data Center Node in the
codes. So the codes generated a exception and could not
reach the codes that add buttons. This patch fixed the problem.

Change-Id: I33802bc7d3a703e6b15b7674a8255a50cf34e4f0
Bug-Url: https://bugzilla.redhat.com/1079180
Signed-off-by: Ma Like <like.ma@cs2c.com.cn>