core: fix reattempt to go to maintenance mechanism
We used to calculate the time from which we should retry to switch the
host to maintenance in case of failure while changing the status of the
host from preparing to maintenance to other status. This is wrong. We
should calculate it when we set the host's status to preparing to
maintenance.

Change-Id: I39835028d5696b9e7a3473852ccf5bf70d3d74d4
Bug-Url: https://bugzilla.redhat.com/1104030
Signed-off-by: Arik Hadas <ahadas@redhat.com>