webadmin: Exception while showing an error message
The regression is caused by
http://gerrit.ovirt.org/#/c/25837
which expected that the new ArrayList(someSize) creates a list of size
someSize. This is not the case so consequently the Collections.fill left the
list empty and consequently the actions.get(actionNum++) failed on an
exception.

Fixed by a simpler and more reliable way of filling a list using a loop and an
add() method.

Change-Id: Ic9714b3b92ff79b3eb12ead5dfb96d22fed41cca
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>