core: Use validation on entites related to one SD.
Remove all the entities which are related to multiple Storage Domains
from the list of Vms/Templates which we validate.

VM/Template with multiple disks on different Storage Domain will not be deleted from the engine
but instead will be kept in the engine without the disks related to it.

Change-Id: I9bc46616a164e6d6ec87fbd4ac67f84823bec4de
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>