core:Force remove should not delete VMs/Tempaltes with MSD.
When the admin chooses to force destroy a storage domain, the operation
should remove all the entities related to the Data Storage Domain.

Since version 3.1 we have added the ability for VMs to have disks on multiple storage domains.
Today, when using force reomve on Storage Domain,
we also remove VMs/Templates with disks on multiple storage domains,
those disks on the other data storage domains are left as floating disks in the system,
even though they might contain snapshots of the deleted VM/Template.

This patch fix this behaviour by leaving the VMs/Templates with disks on MSD,
when using force remove.
It also deletes the devices which are related to those VMs/Templates,
since until now those devices were deleted also with the VM/Template.

Change-Id: Ic222ee55a711ea9101bf20ac812219714f70aab3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1108242
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>