core, engine: Disable Migrate On Error on ppc64 clusters
This change handles the use case where a host enters a cluster without
a CPU name, changing this value to have the same value as the one
reported by the host. If the architecture of the host does not support
migration, this change forces the Migrate On Error option to be
disabled.

Change-Id: I556ef1ce4896ddeb39b93a93520fd0535b4fa28c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1057178
Signed-off-by: Vitor de Lima <vitor.lima@eldorado.org.br>
Signed-off-by: Gustavo Pedrosa <gustavo.pedrosa@eldorado.org.br>