core: [RFE] add support for custom fencing
Up to now if a customer wants to add an additional fencing device, he had
to tweak the oVirt fencing configuration values while those changes were
overridden when engine-setup upgrades oVirt.

This patch adds a custom configuration key for each of the relevant
fencing configuration keys such that the customer can set those keys
using the engin-config utility and the corresponding values are merged
with the original after validating that their formats are valid.

Change-Id: Ie887b9025d0813c4d101a886614364127aea75a0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=878662
Signed-off-by: Eli Mesika <emesika@redhat.com>