webadmin: Vm dialog- bugs with nic naming
New nic name should be- nic + the highest existing nic num + 1.

Since remove from the vnics list didn't work the list contained the old
nics and the naming logic didn't work.

To fix the issue 'vnics' type was changed from Set to List.

VmNetworkInterface is mutable and changing its fields affects hashcode
and equals.
The javadoc for Set says:
Note: Great care must be exercised if mutable objects are used as set
elements. The behavior of a set is not specified if the value of an
object is changed in a manner that affects equals comparisons while the
object is an element in the set.

Change-Id: Ib8594a29ba9c8ed32e4a0d4f8df78212dfc9f40f
Bug-Url: https://bugzilla.redhat.com/1049789
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>