core: ProcessOvfUpdateForStorageDomain - updateDate may be null
When a new ovf store is created for domain, there's no last update date
as the ovf data was never updated to that domain.
To make it obvious that the ovf store was never updated, we can put null
value to indicate to it was never updated rather than excluding the key
from the description.

Change-Id: Id9743b2658d16039865cd3470bf822b5e5e137ff
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1109567
Signed-off-by: Liron Aravot <laravot@redhat.com>