engine: Configure unique host name on neutron.conf
The agent name should be determined by the engine to support its
association with the ml2 port bindings.
The "host" property within the neutron.conf is used to identify
the agent on the neutron node, when the agent communicates with
the neutron node.

When a port is being allocated on the neutron node, the engine should
provide also the agent the port is intended to be run on. The agent is
identified by that "host" property.
By identifying the agent, the ml2 driver resolves the plugin type
(openvswitch, linuxbridge or any other supported plugin).

So the engine should be able to define a unique host name for each
agent, first for the agent which reports its name to the neutron node,
and later-on when ports are allocated on each agent.

Since the host name is unique, and could be changed only by
re-installing the host, it makes sense to use it as the agent name.

Change-Id: I18db7cce8769b23b1fafcf5ac31ebdb7556a7cec
Signed-off-by: Moti Asayag <masayag@redhat.com>