webadmin: data copied from incorrect source
when the custom instance type is selected and the template selection changes,
the data has been copied always from the newly selected template.

This is not always correct. The problem is on edit VM/pool when the VM can
already have some fields changed comparing to the template and the source has
to be the VM/pool itself.

The logic deciding about the appropriate source has already been implemented in
the getSource() method and was used in updateFields() which is supposed to be
called at the end of the initialization of the dialog. They were some race
situations when the updateFields() has been called before the setSelectedItem
on the template causing incorrect data to be copied to the dialog.

Fixed by using the proper getSource() instead of hardcoded usage of the
selected template.

Change-Id: I3f0114e7fb5741694127d5ab50b6953e98c58139
Bug-Url: https://bugzilla.redhat.com/1104318
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>