core: Deadlock detected when performing plug/unplug VNIC action
A previous patch already sorted all updates by the entity ID.
This patch makes sure that VmDeviceUtil.updateBootOrderInVmDeviceAndStoreToDB
stores in a batch a sorted list of the VmDevices, so that no dead lock can occur.
As a side effect, a performance boost for the update occurs, as instead of in a
loop all updates go through batch.

Change-Id: I40e04d943076d046c379d40004157585cb141771
Bug-Url: https://bugzilla.redhat.com/1102689
Signed-off-by: lzelkha@redhat.com <lzelkha@redhat.com>