core: set device to "lun" on VmInfoBuilder
When starting a virtual machine with plugged lun disk using the virtio
interface set the device type to "lun" instead of "disk".

The rationale behind this patch is to provide the same behavior as when
the disk is hotplugged to a running virtual machine: provide the
capability of using generic scsi commands on the device inside the
virtual machine.

Change-Id: I7ceffe36e1145d4783c94f62043dfad694fdd290
Bug-Url: https://bugzilla.redhat.com/1110687
Signed-off-by: Xavi Francisco <xfrancis@redhat.com>