tools : Upgrades from 3.5 on should look for Command Coordinator related changes to Aysc Tasks
From 3.5 on Async Tasks exists in two different
tables, async_tasks and command_entities. When
upgrading from 3.5.x to higher versions we need
to check if there any rows in command_entities
table where callback_enabled is true and
callback_notified is false. If there are any
rows upgrades should not be allowed.

Change-Id: Ib2c240301cd37df195d60fd8949cd45de78f183e
Bug-Url: https://bugzilla.redhat.com/1108162
Signed-off-by: Ravi Nori <rnori@redhat.com>