core: avoid update disk dynamic data gathered from disk snapshot
When a disk snapshot is plugged to the vm, a transient image is created on the
host local storage to allow read/write to that plugged 'disk'.
Therefore when the disk related stats are recieved for a VM that a disk snapshot
is plugged to, they shouldn't be updated in the disc_image_dynamic table.

The update is avoided at the DB level as we are trying to reduce the
number of the queries in VdsUpdateRuntimeInfo (in oppose to load the
device from the db and perform the check in the engine).

Change-Id: Iaa25d040327b4cfe9b049fa2ec2638893190e8ef
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1110830
Signed-off-by: Liron Aravot <laravot@redhat.com>