core: change the order of canDoAction checks on hibernate
We check if the VM is part of pool after checking if the VM is
stateless. Since every VM in pool is stateless, we'll never get to this
check. This patch change the order of these checks, to get the specific
error for the case where the VM is part of pool.

In addition, removing redundant prefix on log printings.

Change-Id: Ie9b98e797ad3f38f19aea45be2f7fe8d1a7635a8
Signed-off-by: Arik Hadas <ahadas@redhat.com>