core: enable <not equal> operator for tag search
This patch allows searching Tags with != operator.
Up to now, when "tag" is used in the search criteria, we had used
StringOnlyEqualConditionRelationAutoCompleter to validate the tag value,
this class allows only for "=" operator.
Actually, there is no reason for this limitation and we can safely
replace this class use with StringConditionRelationAutoCompleter that
allows both "=" and "!=" operators.
This makes StringOnlyEqualConditionRelationAutoCompleter unused and
therefor it was removed.

Change-Id: Icb8fd9c3f75a62841b8d89933c16ac2077c95e4d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1092609
Signed-off-by: Eli Mesika <emesika@redhat.com>