restapi: JAXB provider
Currently we use a custom message body reader to parse XML input
documents, and we use the default message body writer to generate XML
documents. This patch converts the custom message body reader into a
provider that acts as both reader and writer. This reduces the number of
JAXB contexts created and will simplify removing some dependencies on
Resteasy.

Change-Id: I5208c9d41ab892f55c85ef095f25648f1a95d8df
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>