core: Replaced message of GetIsoListError
The getUserRequestForStorageDomainRepoFileList method used to return
null when the storage domain was not valid.
Now it throws a VdcBLLException with GetIsoListError type instead.
Also, an informative message was added to GetIsoListError.
There's no need to catch the exception in 'executeQueryCommand' in
GetImagesListQueryBase or in GetImageByIdQuery because the one who
calls them, QueriesCommandBase, does the job.

Change-Id: I3e6dc6fbe03ffba94cff8821b35eacf062e42362
Bug-Url: https://bugzilla.redhat.com/918949
Signed-off-by: Idan Shaby <ishaby@redhat.com>