core: CDI workaournd jboss 7.1.1 bug
To workaround jboss 7.1.1 bug the Injector bean must be moved to BLL
module so it will be injected into Backend.java and initialize on
startup.

Till master will support wildfly or at least jboss version EAP 6.x or
greater CDI beans can't be detected across modules.

To overcome that one must create a producer method and put it under
BllCDIAdapter and there to create or use a getInstance incase its a
singleton

Change-Id: If996ccbce18fcc5471acff49f5bef8c70269e0aa
Signed-off-by: Roy Golan <rgolan@redhat.com>