core: avoid deadlocks on images related updated
Currently VdsUpdateRuntimeInfo updates the disc_image_dynamic table
according to the statistics retreived from the running vms.
Those updates are executed in a batchUpdate call, which currently
executes all the updates within the same transaction, as there is a FK
between the disk_image_dynamic on postgres version earlier than 9.3 a
lock will be taken on the referenced row which means that any other
concurrent operation on those images may cause to a deadlock unless they
are sorted.

Sorting the images on all the flows in more tricky and unneeded as this
issue has been resolved in postgres 9.3
(http://www.postgresql.org/docs/9.3/static/release-9-3.html) in the
meanwhile we can drop the fk constraint and delete the records by adding
trigger.

Change-Id: Ifa160402787a7902169ab5fd3ce6661918fdebb2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1112720
Signed-off-by: Liron Aravot <laravot@redhat.com>