engine : NPE while using reflection for AsyncTask type
If the server is restarted before the async task place holder
can be updated with the vdsm id, the async task manager
creates a new async task row with action_type unknown. If
a command is constructed from this data for CommandAsyncTask
VdcActionType Unknown results in no command being found throwing
an exception.

When action_type is unknown, SPMAsync task is constructed instead
of command async task.

Change-Id: Ic186992e1f1d4d6a870b323cc018b6cf3e67645a
Bug-Url: https://bugzilla.redhat.com/1112557
Signed-off-by: Ravi Nori <rnori@redhat.com>