core: allow successful retries of Live Merge
Change RemoveSnapshotSingleDiskLiveCommand and related commands to
allow for successful retries after deletion or merge fails.

Upon merge failure, images are marked illegal but remain in the volume
chain.  Retries will start over and attempt to merge the images again.

Upon deletion failure (but after a successful merge), the volume chain
is updated to reflect the new chain; however, the orphaned images are
re-associated with the snapshot to be deleted, marked illegal, and set
to have no parents or children.  Upon retry, the command will detect
this state and skip to the DESTROY_IMAGE execution step.

Change-Id: Ib794822c6b03cb05e00dc7e99113127aaed43ce9
Signed-off-by: Greg Padgett <gpadgett@redhat.com>