webadmin: Fixed a small mess in host network popup
Probably due to a lapse in concentration at the time, the
changeability of network custom properties depended upon the ability
to change boot protocol, where it should have depended solely on
whether the network was marked to by synchronized (it should still
work when the boot protocol is set to "None", which disables the boot
protocol fields).

Also, while fixing this I noticed that host QoS wasn't disabled for
the same cases, when it actually should be.

Change-Id: I486cc099cf7698f0d685c398c00215b0bab2f3fa
Signed-off-by: Lior Vernia <lvernia@redhat.com>