core: enable hyperv optimization on windows
On VDSM side, we are catching-up with HyperV
enligthenment support and recommended settings.

Some of them are grouped under the 'hypervEnable'
toggle and it is required to be explicitely enable
for various reasons, the first one is they may be not
completely transparent to not-windows guests.

This patch automaticcaly enables these optimizations
for windows guests.

Change-Id: I01d50a539347211a6729cfb86922d60a6244d495
Bug-Url: https://bugzilla.redhat.com/1110305
Relates-To: https://bugzilla.redhat.com/1107835
Relates-To: https://bugzilla.redhat.com/1083529
Signed-off-by: Francesco Romani <fromani@redhat.com>