restapi: Deploy using modules
Currently the RESTAPI application is deployed as a nested .war file
inside the .ear of the engine, actually as two .war files, one for /api,
and another for /ovirt-engine/api. This double deployment causes class
cast exceptions because classes are loaded twice. To avoid this issues
this patch changes the RESTAPI application so that it is deployed
outside of the .ear and users JBoss modules for all its components, so
that no class will be loaded twice.

Change-Id: Ie4079277f60e5521d9b61379fa78713a5c4ccc10
Bug-Url: https://bugzilla.redhat.com/1113485
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>