core: Adding support for detach storage from old pool
Adding a parameter for indication when we want to detach a storage
domain from its old pool in the meta data.

The vdsm action forcedDetachStorageDomain has been changed to indicate
that only domain with empty GUID pool id, will be forced detached in the
meta data.
This change in the engine is related to this change

Change-Id: Ieab6d838778246c2d51b7b77ba655275016148df
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>