core: handle fence agent power wait param on stop
When a host restart is dome manually or as a result of a
non-responsive host treatment and in the case that power wait parameter
is used the host may stay in 'off' state and even release its lock on HA
VMs before the host is really down

This is the scenario:
1) A restart command is issued and actually performed as
   stop -> wait for 'off' status -> start -> wait for 'on' status
2) power wait parameter is added to the command implicitly or explicitly
giving a delay of X seconds before the operation is actually performed
3) The fence agent script returns immediately with 'off' status
4) A 'on' command is sent to the fence agent by the start operation
5) X seconds passed and the host is actually shutdown

This patch handles this by adding a new configuration value that maps
fence agents to the name of the parameter for power wait and it also
adds a custom configuration value that may be set for custom fence
agents.

Upon stop operation, we will wait a fixed delay (5 seconds hard-coded)
before starting to sample the host status, if power wait parameter is
used, the value is extracted and we will wait for 5 + X seconds.

Change-Id: I310e076ecf84988cacd0b179954d2460d7988b91
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1093742
Signed-off-by: Eli Mesika <emesika@redhat.com>